+++ packages/pangea-node-sdk/tests/integration/redact.test.ts
@@ -2,9 +2,10 @@ import PangeaConfig from "../../src/config";
 import { PangeaErrors } from "../../src/errors";
 import RedactService from "../../src/services/redact";
 import { it, expect } from "@jest/globals";
+import { TestEnvironment, getTestDomain, getTestToken } from "../../src/utils/utils";
 
-const token = process.env.PANGEA_INTEGRATION_TOKEN || "";
-const testHost = process.env.PANGEA_INTEGRATION_DOMAIN || "";
+const token = getTestToken(TestEnvironment.LIVE);
+const testHost = getTestDomain(TestEnvironment.LIVE);
 const config = new PangeaConfig({ domain: testHost });
 const redact = new RedactService(token, config);